moved hc_rook_permissions_manager methods to hc_rook_health_permissions_manager to ensure parity with apple health sdk
