+++ packages/rook_sdk_health_connect/lib/src/hc_rook_yesterday_sync_permissions.dart
@@ -1,6 +1,8 @@
 import 'package:rook_sdk_health_connect/src/platform/rook_sdk_health_connect_platform_interface.dart';
 
-@Deprecated("Methods from HCRookPermissionsManager offer better functionality.")
+@Deprecated(
+  "Methods from HCRookHealthPermissionsManager offer better functionality.",
+)
 class HCRookYesterdaySyncPermissions {
   HCRookYesterdaySyncPermissions._();
 
@@ -16,7 +18,8 @@ class HCRookYesterdaySyncPermissions {
   ///
   /// @return True if the permissions are granted, false otherwise.
   @Deprecated(
-      "Methods from HCRookPermissionsManager offer better functionality. Replace with HCRookPermissionsManager.checkAndroidPermissions")
+    "Replace with HCRookHealthPermissionsManager.checkAndroidPermissions",
+  )
   static Future<bool> hasAndroidPermissions() {
     return RookSdkHealthConnectPlatform.instance
         .hasYesterdaySyncAndroidPermissions();
@@ -32,7 +35,8 @@ class HCRookYesterdaySyncPermissions {
   /// • FOREGROUND_SERVICE_HEALTH (API Level 34+)
   /// ```
   @Deprecated(
-      "Methods from HCRookPermissionsManager offer better functionality. Replace with HCRookPermissionsManager.requestAndroidPermissions")
+    "Replace with HCRookHealthPermissionsManager.requestAndroidPermissions",
+  )
   static Future<void> requestAndroidPermissions() {
     return RookSdkHealthConnectPlatform.instance
         .requestYesterdaySyncAndroidPermissions();
@@ -50,7 +54,8 @@ class HCRookYesterdaySyncPermissions {
   ///
   /// @return True if the permissions are granted, false otherwise.
   @Deprecated(
-      "Methods from HCRookPermissionsManager offer better functionality. Replace with HCRookPermissionsManager.checkHealthConnectPermissions")
+    "Replace with HCRookHealthPermissionsManager.checkHealthConnectPermissions",
+  )
   static Future<bool> hasHealthConnectPermissions() {
     return RookSdkHealthConnectPlatform.instance
         .hasYesterdaySyncHealthConnectPermissions();
@@ -66,7 +71,8 @@ class HCRookYesterdaySyncPermissions {
   /// • BODY
   /// ```
   @Deprecated(
-      "Methods from HCRookPermissionsManager offer better functionality. Replace with HCRookPermissionsManager.requestHealthConnectPermissions")
+    "Replace with HCRookHealthPermissionsManager.requestHealthConnectPermissions",
+  )
   static Future<void> requestHealthConnectPermissions() {
     return RookSdkHealthConnectPlatform.instance
         .requestYesterdaySyncHealthConnectPermissions();
